Display Objects, Camera management, Shaders Generator (https://github.com/enso-org/ide/pull/89)
Original commit: https://github.com/enso-org/ide/commit/5c37d3c9bc40dbf866c5970fb5d97a7db003566d

+++ gui/lib/system/web/src/resize_observer.rs
@@ -6,7 +6,7 @@ use wasm_bindgen::JsValue;
 // === Types ===
 // =============
 
-pub type Listener = Closure<dyn FnMut(i32, i32)>;
+pub type Listener = Closure<dyn FnMut(f64,f64)>;
 
 // ===================
 // === JS Bindings ===
